payment_links_service/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
pytest==8.3.4
requests==2.32.3